// src/types/global.types.ts
export type DrillType =
  | "A7"
  | "A8"
  | "A9"
  | "A10"
  | "C4"
  | "C5"
  | "C11"
  | "C14"
  | "SK1"
  | "SK2"
  | "SK3"
  | "SK6"
  | "S12"
  | "S13";

export type StatusNumbers = 217 | 218 | 219 | 216 | 220 | 215;
export type Status =
  | "Готовност"
  | "СВАЛЕН ОТ УПОТРЕБА"
  | "Изчакване"
  | "Закъснение"
  | "Смяна на смени"
  | "*";

export type MonthBG =
  | "Януари"
  | "Февруари"
  | "Март"
  | "Април"
  | "Май"
  | "Юни"
  | "Юли"
  | "Август"
  | "Септември"
  | "Октомври"
  | "Ноември"
  | "Декември";

export const monthNamesBG: Record<number, MonthBG> = {
  1: "Януари",
  2: "Февруари",
  3: "Март",
  4: "Април",
  5: "Май",
  6: "Юни",
  7: "Юли",
  8: "Август",
  9: "Септември",
  10: "Октомври",
  11: "Ноември",
  12: "Декември",
};

export type ShiftNames = "Първа" | "Обяд" | "Втора" | "Нощна";